apparently this really solves the Next dark magic

+++ pages/api/setup.tsx
@@ -36,6 +36,7 @@ export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse,
 ) {
+  if (req.method == "OPTIONS") return res.status(200).json({ message: "ok" })
   if (req.method == "DELETE") {
     await deleteAllData()
     res.json({
